Made the colorMode persist through the pages and initial it based on the user system settings
What I have done: I found a way to initale the colorMode based on the user system settings and after that store the current mode in a localeStorage variable. By doing this I was able to make the color mode persist after switching the .html

What I learned: I figured out how to accsess the value of a :root variable and I found out locale Storage exist and how to set and get Items in it.

Current Challenge: -

Next step: I want to create a nice button to toggle mode